Setup and configure mkdocs, write initial docs
We need a place to write documentation for PineSAM and accept
contributions from non-technical end users who wish to add items to the
FAQ, add information, and otherwise grow the project.

This is a fairly large PR but it is the result of about a week of work
getting things setup and configured how we think it should be.